QA_by_using_Google_palm/requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
google-generativeai==0.2.2
InstructorEmbedding==1.0.1
torch